cave update, instead downloaded libnoise through vcpkg

+++ gamerenderer.cpp
@@ -59,7 +59,7 @@ bool GameRenderer::drawChunk(WorldChunk& p_chunk, GameWindow& p_window) {
 	int w, h;
 	SDL_GetWindowSize(p_window.window, &w, &h);
 	cam.setDimensions(aspect);
-	glm::mat4 finalTransform = cam.getTransformMat4(dm.w, dm.h, w, h);
+	glm::mat4 finalTransform = cam.getTransformMat4((float)dm.w, (float)dm.h, (float)w, (float)h);
 	imageShader.setMat4Uniform("transform", finalTransform);
 
 	glDrawArrays(GL_TRIANGLES, 0, p_chunk.getVBOSize());
